refactor: Move container-definitions errors into core-definitions in order to eliminate container-utils (#16868)
[AB#5073](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5073)

We are working to remove the @fluidframework/container-utils package, as
we do not intend / expect external consumers to use it.

As a part of this effort, we have identified that there are two types of
definitions within @fluidframework/container-utils/errors: 1) classes
from "@fluidframework/container-definitions" and 2) classes from
"@fluidframework/telemetry-utils". 1) is only there becuase 2 depend on
them. However, "@fluidframework/telemetry-utils" shouldn't depend on and
pull in "@fluidframework/container-definitions" since they should be
able to be used anywhere in the stack without depending on container
packages. So we are migrating both 1) and 2) into
"@fluidframework/core-interfaces".

This PR makes this move and updates internal uses of the
"@fluidframework/container-definitions" to pull from
"@fluidframework/core-interfaces".

## Partner Repos
Loop uses: ContainerErrorType, IErrorBase, IThrottlingWarning from
container-definitions.
Whiteboard uses: ContainerErrorType, IErrorBase from
container-definitions.
PowerPoint Online doesn't use any of these.
WordJs uses: ContainerErrorType from container-definitions. 
Bohemia uses: ContainerErrorType, IErrorBase, IThrottlingWarning from
container-definitions.

So we'll need to create PRs in these repos to go along with this change.

---------
