+++ packages/common/container-definitions/src/error.ts
@@ -3,10 +3,25 @@
  * Licensed under the MIT License.
  */
 
-import { ITelemetryProperties } from "@fluidframework/core-interfaces";
+import { FluidErrorTypes, IErrorBase } from "@fluidframework/core-interfaces";
 
 /**
- * Different error types the Container may report out to the Host
+ * Different error types the ClientSession may report out to the Host.
+ */
+export const ContainerErrorTypes = {
+	...FluidErrorTypes,
+	/**
+	 * Error indicating an client session has expired. Currently this only happens when GC is allowed on a document and
+	 * aids in safely deleting unused objects.
+	 */
+	clientSessionExpiredError: "clientSessionExpiredError",
+} as const;
+export type ContainerErrorTypes = typeof ContainerErrorTypes[keyof typeof ContainerErrorTypes];
+
+/**
+ * Different error types the Container may report out to the Host.
+ *
+ * @deprecated ContainerErrorType is being deprecated as a public export. Please use {@link ContainerErrorTypes#clientSessionExpiredError} instead.
  */
 export enum ContainerErrorType {
 	/**
@@ -42,35 +57,6 @@ export enum ContainerErrorType {
 }
 
 /**
- * Base interface for all errors and warnings at container level
- */
-export interface IErrorBase extends Partial<Error> {
-	/** errorType is a union of error types from
-	 * - container
-	 * - runtime
-	 * - drivers
-	 */
-	readonly errorType: string;
-
-	/**
-	 * See Error.message
-	 * Privacy Note - This is a freeform string that we may not control in all cases (e.g. a dependency throws an error)
-	 * If there are known cases where this contains privacy-sensitive data it will be tagged and included in the result
-	 * of getTelemetryProperties. When logging, consider fetching it that way rather than straight from this field.
-	 */
-	readonly message: string;
-	/** See Error.name */
-	readonly name?: string;
-	/** See Error.stack */
-	readonly stack?: string;
-	/**
-	 * Returns all properties of this error object that are either safe to log
-	 * or explicitly tagged as containing privacy-sensitive data.
-	 */
-	getTelemetryProperties?(): ITelemetryProperties;
-}
-
-/**
  * Represents warnings raised on container.
  */
 export interface ContainerWarning extends IErrorBase {
@@ -89,7 +75,7 @@ export interface ContainerWarning extends IErrorBase {
  *
  * The following are commonly thrown error types, but `errorType` could be any string.
  *
- * - {@link @fluidframework/container-definitions#ContainerErrorType}
+ * - {@link @fluidframework/core-interfaces#ContainerErrorType}
  *
  * - {@link @fluidframework/driver-definitions#DriverErrorType}
  *
@@ -99,26 +85,3 @@ export interface ContainerWarning extends IErrorBase {
  *
  */
 export type ICriticalContainerError = IErrorBase;
-
-/**
- * Generic wrapper for an unrecognized/uncategorized error object
- */
-export interface IGenericError extends IErrorBase {
-	readonly errorType: ContainerErrorType.genericError;
-	error?: any;
-}
-
-/**
- * Error indicating an API is being used improperly resulting in an invalid operation.
- */
-export interface IUsageError extends IErrorBase {
-	readonly errorType: ContainerErrorType.usageError;
-}
-
-/**
- * Warning emitted when requests to storage are being throttled
- */
-export interface IThrottlingWarning extends IErrorBase {
-	readonly errorType: ContainerErrorType.throttlingError;
-	readonly retryAfterSeconds: number;
-}
