+++ packages/common/container-definitions/src/index.ts
@@ -25,13 +25,10 @@ export {
 	ReadOnlyInfo,
 } from "./deltas";
 export {
+	ContainerErrorTypes,
 	ContainerErrorType,
 	ContainerWarning,
 	ICriticalContainerError,
-	IErrorBase,
-	IGenericError,
-	IUsageError,
-	IThrottlingWarning,
 } from "./error";
 export {
 	ConnectionState,
@@ -70,3 +67,22 @@ export {
 	IRuntime,
 	IRuntimeFactory,
 } from "./runtime";
+
+export {
+	/**
+	 * @deprecated IErrorBase is being deprecated as a public export is moving to "core-interfaces".
+	 */
+	IErrorBase,
+	/**
+	 * @deprecated IGenericError is being deprecated as a public export is moving to "core-interfaces".
+	 */
+	IGenericError,
+	/**
+	 * @deprecated IThrottlingWarning is being deprecated as a public export is moving to "core-interfaces".
+	 */
+	IThrottlingWarning,
+	/**
+	 * @deprecated IUsageError is being deprecated as a public export is moving to "core-interfaces".
+	 */
+	IUsageError,
+} from "@fluidframework/core-interfaces";
