+++ packages/common/core-interfaces/src/index.ts
@@ -3,6 +3,16 @@
  * Licensed under the MIT License.
  */
 
+export type { IDisposable } from "./disposable";
+
+export {
+	FluidErrorTypes,
+	IErrorBase,
+	IGenericError,
+	IUsageError,
+	IThrottlingWarning,
+} from "./error";
+
 export type {
 	ExtendEventProvider,
 	IErrorEvent,
@@ -20,6 +30,17 @@ export {
 	IFluidRunnable,
 	IProvideFluidRunnable,
 } from "./fluidLoadable";
+
+export {
+	IFluidPackageEnvironment,
+	IFluidPackage,
+	isFluidPackage,
+	IFluidCodeDetailsConfig,
+	IFluidCodeDetails,
+	isFluidCodeDetails,
+	IFluidCodeDetailsComparer,
+	IProvideFluidCodeDetailsComparer,
+} from "./fluidPackage";
 
 // Typescript forgets the index signature when customers augment IRequestHeader if we export *.
 // So we export the explicit members as a workaround:
@@ -39,19 +60,6 @@ export {
 	IProvideFluidHandle,
 } from "./handles";
 
-export {
-	IFluidPackageEnvironment,
-	IFluidPackage,
-	isFluidPackage,
-	IFluidCodeDetailsConfig,
-	IFluidCodeDetails,
-	isFluidCodeDetails,
-	IFluidCodeDetailsComparer,
-	IProvideFluidCodeDetailsComparer,
-} from "./fluidPackage";
-
-export { FluidObjectProviderKeys, FluidObject, FluidObjectKeys } from "./provider";
-export type { IDisposable } from "./disposable";
 export type {
 	ILoggingError,
 	ITaggedTelemetryPropertyType,
@@ -65,3 +73,5 @@ export type {
 	TelemetryEventCategory,
 	TelemetryEventPropertyType,
 } from "./logger";
+
+export { FluidObjectProviderKeys, FluidObject, FluidObjectKeys } from "./provider";
